requirements.txt
flask-cors
matplotlib
Flask==2.1.1 
werkzeug==2.1.1

torch==1.10.0
torchvision==0.11.1
opencv-python==4.5.4.58
numpy==1.21.2
Pillow==8.3.2

